+++ src/Contracts/PaymentManager.php
@@ -1,26 +1,26 @@
 <?php
 
-namespace rkujawa\LaravelPaymentGateway\Interfaces;
+namespace rkujawa\LaravelPaymentGateway\Contracts;
 
 interface PaymentManager
 {
     /**
-     * @param \rkujawa\LaravelPaymentGateway\Interfaces\BillableContract $billable
+     * @param \rkujawa\LaravelPaymentGateway\Contracts\BillableContract $billable
      * @param array|mixed $data
-     * @return \rkujawa\LaravelPaymentGateway\Interfaces\PaymentManagerResponse
+     * @return \rkujawa\LaravelPaymentGateway\Contracts\PaymentManagerResponse
      */
     public function createPaymentMethod($billable, $data);
 
     /**
      * @param \rkujawa\LaravelPaymentGateway\Models\PaymentMethod $paymentMethod
      * @param array|mixed $data
-     * @return \rkujawa\LaravelPaymentGateway\Interfaces\PaymentManagerResponse
+     * @return \rkujawa\LaravelPaymentGateway\Contracts\PaymentManagerResponse
      */
     public function updatePaymentMethod($paymentMethod, $data);
     
     /**
      * @param \rkujawa\LaravelPaymentGateway\Models\PaymentMethod $paymentMethod
-     * @return \rkujawa\LaravelPaymentGateway\Interfaces\PaymentManagerResponse
+     * @return \rkujawa\LaravelPaymentGateway\Contracts\PaymentManagerResponse
      */
     public function deletePaymentMethod($paymentMethod);
 }
